+++ src/net_processing.cpp
@@ -29,6 +29,7 @@
 #include <util/system.h>
 #include <util/moneystr.h>
 #include <util/strencodings.h>
+#include <util/validation.h>
 
 #include <memory>
