+++ src/rpc/mining.cpp
@@ -21,8 +21,10 @@
 #include <rpc/util.h>
 #include <shutdown.h>
 #include <txmempool.h>
+#include <util/fees.h>
 #include <util/strencodings.h>
 #include <util/system.h>
+#include <util/validation.h>
 #include <validation.h>
 #include <validationinterface.h>
 #include <versionbitsinfo.h>
