+++ src/ui_interface.h
@@ -129,10 +129,6 @@ void InitWarning(const std::string& str);
 /** Show error message **/
 bool InitError(const std::string& str);
 
-std::string AmountHighWarn(const std::string& optname);
-
-std::string AmountErrMsg(const char* const optname, const std::string& strValue);
-
 extern CClientUIInterface uiInterface;
 
 #endif // BITCOIN_UI_INTERFACE_H
